Moved toProperCase logic moved from Util class to StringFormatter class. Moved business hours button from the CreateRestaurant/EditRestaurant layouts to the manage restaurant layout.
