+++ app/src/main/java/com/gmail/dleemcewen/tandemfieri/Adapters/ManageRestaurantExpandableListAdapter.java
@@ -254,6 +254,7 @@ public class ManageRestaurantExpandableListAdapter extends BaseExpandableListAda
         Button manageMenuItems = (Button)convertView.findViewById(R.id.manageMenuItems);
         Button viewSales = (Button)convertView.findViewById(R.id.viewSales);
         Button viewDeliveryArea = (Button)convertView.findViewById(R.id.viewDeliveryArea);
+        Button setBusinessHours = (Button)convertView.findViewById(R.id.businessHours);
         Button manageDrivers = (Button)convertView.findViewById(R.id.manageDrivers);
         Button rateDrivers = (Button)convertView.findViewById(R.id.rateDrivers);
 
@@ -280,6 +281,12 @@ public class ManageRestaurantExpandableListAdapter extends BaseExpandableListAda
                 Toast
                     .makeText(context, "Viewing sales for " + selectedChild.getName(), Toast.LENGTH_SHORT)
                     .show();
+            }
+        });
+        setBusinessHours.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Toast.makeText(context, "Opening soon!", Toast.LENGTH_SHORT).show();
             }
         });
         viewDeliveryArea.setOnClickListener(new View.OnClickListener() {
