+++ app/src/main/java/com/gmail/dleemcewen/tandemfieri/EditRestaurantActivity.java
@@ -17,6 +17,7 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import com.gmail.dleemcewen.tandemfieri.Entities.Restaurant;
+import com.gmail.dleemcewen.tandemfieri.Formatters.StringFormatter;
 import com.gmail.dleemcewen.tandemfieri.Logging.LogWriter;
 import com.gmail.dleemcewen.tandemfieri.Repositories.Restaurants;
 import com.gmail.dleemcewen.tandemfieri.Utility.Util;
@@ -46,7 +47,6 @@ public class EditRestaurantActivity extends AppCompatActivity implements Adapter
     private EditText zipCode;
     private EditText deliveryCharge;
     private String state;
-    private Button businessHours;
     private Button updateRestaurant;
     private Button cancelUpdateRestaurant;
 
@@ -90,7 +90,6 @@ public class EditRestaurantActivity extends AppCompatActivity implements Adapter
         states = (Spinner)findViewById(R.id.state);
         zipCode = (EditText)findViewById(R.id.zipcode);
         deliveryCharge = (EditText)findViewById(R.id.deliveryCharge);
-        businessHours = (Button)findViewById(R.id.businessHours);
         updateRestaurant = (Button)findViewById(R.id.createRestaurant);
         cancelUpdateRestaurant = (Button)findViewById(R.id.cancelRestaurant);
     }
@@ -100,13 +99,6 @@ public class EditRestaurantActivity extends AppCompatActivity implements Adapter
      */
     private void bindEventHandlers() {
         states.setOnItemSelectedListener(this);
-
-        businessHours.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Toast.makeText(EditRestaurantActivity.this, "Opening soon!", Toast.LENGTH_SHORT).show();
-            }
-        });
 
         updateRestaurant.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -192,7 +184,7 @@ public class EditRestaurantActivity extends AppCompatActivity implements Adapter
 
         // Find the user's state in the array of states
         String[] arrayOfStates = getResources().getStringArray(R.array.states);
-        int positionOfUserState = Arrays.asList(arrayOfStates).indexOf(Util.toProperCase(restaurant.getState()));
+        int positionOfUserState = Arrays.asList(arrayOfStates).indexOf(StringFormatter.toProperCase(restaurant.getState()));
         states.setSelection(positionOfUserState);
 
         //set button text
