+++ app/src/main/res/layout/activity_create_restaurant.xml
@@ -104,24 +104,15 @@
         android:layout_alignParentStart="true"
         android:id="@+id/deliveryCharge"
         android:hint="@string/hintDeliveryCharge" />
-    <Button
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:textSize="@dimen/button_text_size"
-        android:id="@+id/businessHours"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentLeft="true"
-        android:layout_below="@id/deliveryCharge"
-        android:text="@string/businessHours" />
     <LinearLayout
         android:id="@+id/actionsContainer"
-        android:layout_marginTop="20dp"
+        android:layout_marginTop="64dp"
         android:layout_width="match_parent"
         android:layout_height="64dp"
         android:orientation="horizontal"
         android:layout_alignParentStart="true"
         android:layout_alignParentLeft="true"
-        android:layout_below="@id/businessHours">
+        android:layout_below="@id/deliveryCharge">
         <Button
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
